Remove useFlow API which accepts a regular Flow
Summary:
This diff removes the API that accepts a regular `Flow`. This means that only `StateFlow` is supported.

In a multithreaded scenario, when Component's `render()` method is invoked on a background thread there is a window in which a value emitted via regular `Flow` can be lost. This is because we always `collect{}` the `Flow` on the main thread, so if the `Flow` is updated between `useFlow` call and `collect{}`, then such value will be lost.

`StateFlow` doesn't have this issue because as stated in the [docs](https://kotlinlang.org/api/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines.flow/-state-flow/): `"State flow always has an initial value, replays one most recent value to new subscribers, ..."`.

The same is true for [LiveData](https://developer.android.com/reference/androidx/lifecycle/LiveData#observe(androidx.lifecycle.LifecycleOwner,androidx.lifecycle.Observer%3C?%20super%20T%3E)) - `"Adds the given observer to the observers list within the lifespan of the given owner. The events are dispatched on the main thread. If LiveData already has data set, it will be delivered to the observer."`

A regular `Flow` doesn't remember any values so it needs to be converted to `StateFlow` before passing it to Litho.

Reviewed By: astreet

Differential Revision: D49417629

fbshipit-source-id: d9f81332ee87478c07fd17c91bca3b70ebc8883e