Fix UIDebugger bug with nested tree holder
Summary:
A User reported issue with this screen where uidebugger hangs. https://fb.workplace.com/groups/443457641253219/permalink/1700645907082727/

It was due to the UIDebugger seeing a node with a circular reference to itself, ie its parent id was itself causing an infinite loop while walking up the tree.

While digging i noticed that this occured with nested tree holder, also i noticed a missing return statement that looked suspicious This seems to fix the UIDebugger issue.

While this is a unsatisfactory understanding of the issue it seems completely sane and fixes the bug

Reviewed By: adityasharat

Differential Revision: D50093969

fbshipit-source-id: a397bf5bf8250822d27435dafbda74c7709d1303