Add errata for fix to marginStart/End for row-reverse flex direction
Summary:
This stack is ultimately aiming to solve https://github.com/facebook/yoga/issues/1208

This adds an value to the Errata enum. I will use this to gate this fix as there is potential for users to rely on this bug or have a hack in place to fix it and this would be a breaking change.

Differential Revision: D50145273

fbshipit-source-id: 2402f371acae7f7ec2b75835be1b9fe7844b77ec