Revert D49766621: Multisect successfully blamed "D49766621: [Litho][Layout Caching]Remove legacy code to ship layout caching" for test or build failures
Summary:
This diff is reverting D49766621
Depends on D50237375
D49766621: [Litho][Layout Caching]Remove legacy code to ship layout caching by pentiumao has been identified to be causing the following test or build failures:

Tests affected:
- [automation_assistant_sdk_sample_debug](https://www.internalfb.com/intern/test/562949987519956/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3268977
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: pentiumao

Differential Revision: D50237397

fbshipit-source-id: bd771f870839f2a4341dc661b36e9dfc57513a9f