Fix excluding from IM in Primitive API
Summary:
Prior to this change, using `shouldExcludeFromIncrementalMount` in Primitive API doesn't always work correctly. Specifically it won't work if `shouldExcludeFromIncrementalMount` is set in `MountConfigurationScope`. For example:

```
render() {
  // this would work
  shouldExcludeFromIncrementalMount = true

  Primitive(
    layoutBehavior = ...
    mountBehavior = MountBehavior {
      // this wouldn't work
      shouldExcludeFromIncrementalMount = true
    }
  )
}
```

The reason is that `MountConfigurationScope` logic is executed when the underlying `RenderUnit` is created. And prior to this diff, the RU was created lazily when `primitive.renderUnit` was called for the first time. Unfortunately in `PrimitiveComponent`, we're checking `shouldExcludeFromIncrementalMount` flag before accessing `primitive.renderUnit`.
So if `shouldExcludeFromIncrementalMount` flag was set within `MountBehavior{}` block then Litho would check the value of `shouldExcludeFromIncrementalMount` before `MountBehavior{}` is executed.

The fix is to create the underlying RU and thus initialize `shouldExcludeFromIncrementalMount` at the time when `MountBehavior` object is created. It doesn't make sense to do it lazily because that RU needs to be created during the resolve phase anyways.

Reviewed By: adityasharat, fabiocarballo

Differential Revision: D50217627

fbshipit-source-id: ade7e6acdc76b73ab84e40cf81a99df6a8ef10f9