Add ability to specify `DelegateBinder` as the custom binder.
Summary:
While attempting to add yet another custom binder, I have noticed that with the current implementation it is not possible to use `shouldUpdate` taking account specific previous/new data. To accomodate that I'm adding the ability to specify a `DelegateBinder` instead which provides this mechanism out of the box.

Currently I'm duplicating this because I plan to put it behind a QE since critical logging code on Feed and Reels is using this.

This change will allow us to do things like (which can be seen in the next diff of the stack):

```

fun Style.commentRowViewpointTrackingBinder(
    commentModel: CommentListRowModel,
    viewLogger: CommentViewLayerLogger
): Style =
    delegateViewBinder(
        DelegateBinder.createDelegateBinder(
            commentModel, CommentRowViewpointTrackingBinder(commentModel, viewLogger)))

internal class CommentRowViewpointTrackingBinder(
    private val commentModel: CommentListRowModel,
    private val viewLogger: CommentViewLayerLogger,
) : RenderUnit.Binder<CommentListRowModel, View, Any?> {

  override fun shouldUpdate(
      currentModel: CommentListRowModel,
      newModel: CommentListRowModel,
      currentLayoutData: Any?,
      nextLayoutData: Any?
  ): Boolean {
    return currentModel.id != newModel.id
  }

  override fun bind(
      context: Context?,
      content: View,
      model: CommentListRowModel?,
      layoutData: Any?
  ): Any? {
    viewLogger.registerRowToViewPoint(content, commentModel)
    return null
  }

  override fun unbind(
      context: Context?,
      content: View,
      model: CommentListRowModel?,
      layoutData: Any?,
      bindData: Any?
  ) {
    viewLogger.unregisterRowToViewPoint(content)
  }
}
```

Reviewed By: pentiumao

Differential Revision: D50364937

fbshipit-source-id: 37e12d92a1973ba352d08a681cc1e36ecbc3a591