Makes EventHandler::dispatchEvent return result
Summary:
Several place throughout the code, and in generated code event are not dispatched directly using the event handler, instead the dispatch info, and internal dispatcher are manually accessed to to dispatch the event. This is primarily done to handle cases where the event returns a result.

This is a problem because we are trying to log event dispatches in `EventHandler::dispatchEvent` so this diff makes `EventHandler::dispatchEvent` return the result to eventually remove access to the dispatcher.

Reviewed By: astreet

Differential Revision: D50167802

fbshipit-source-id: 6e22369585d167c3a1d77d2ea3609db1ae47eb7f