clean up RenderResult
Summary:
A `RenderResult` should be the output of laying out and reducing a resolved node tree. It should not be concerned with resolving a `ResolveFunc`. A `ResolveFunc` should be handled by a top-level integration (e.g `RenderState` or `ComponentTree`) with the ability to consume and apply state updates.

With this diff, `RenderResult.render(...)` will only accept a resolved node tree.

Reviewed By: Enfield

Differential Revision: D50271378

fbshipit-source-id: c3caad118ebdb683f203a25a63356e0248105f7b