remove layout data check to unbreak Litho layout caching
Summary:
Litho makes assumptions around layout data that are not completely correct. Unfortunately the layout data check in `MountState` breaks Litho's layout caching. I'm reverting the check to unblock experiments on layout caching. We'll reland the check with more soak time to ensure that we catch any layout caching issues in Litho.

I'll try and get a pick for this between today and tomorrow

Reviewed By: pentiumao

Differential Revision: D50702380

fbshipit-source-id: e080e032d01097c864d4f23141a7a4c33207ad57