Fix missing size check for cached result
Summary: This is to fix the bug that was found in live video where a fixed size Recycler wasn't rendered correctly by rotate device to landscape mode. After my investigation, this was due to we missed a corner case for `onBoundsDefined` that a fixed size component could skip yoga measurement with a different size.

Reviewed By: adityasharat

Differential Revision: D50777075

fbshipit-source-id: 9a1b041342439b4241d1f362fe59757597c4613d