Remove OSS Yoga source dirsync and build (#963)
Summary:
Litho in OSS uses published Yoga binaries for releases (see D6246144) but has its own Yoga source build used when building for development.

I don't have a clear understanding of whether anyone owns the Litho OSS NDK build, but its been failing for the last few months after I added some C++ 14 usage (we were already using C++ 17 as well, but the Litho OSS build was set to not emit any warnings). I reassigned an OSS CI failure to Litho OSS oncall to take a look at the Litho CMake setup, but it was auto closed and never acted on.

Dirsyncs are not free. E.g. merge conflict resolution is handled for each separate copy of a file. This also means one more separate Yoga build to need to maintain (we have too many already).

This change moves the Litho OSS build exclusively to the Yoga AAR published to Maven Central, and removes NDK usage from the Litho OSS build. It also bumps `minSdk` to 21, to match that of the Yoga package we already released with (that is also the new min for NDK 26).

This means the source is no longer exactly in sync. This is a non-issue in practice, since we are already bumping the release binary so rarely, and Litho hasn't been taking much advantage of recent new APIs. Litho snapshot builds could also be set up to rely on Yoga snapshot builds, to keep those more in sync.

More context in https://fb.workplace.com/groups/litho.support/permalink/3639932409661434/

Pull Request resolved: https://github.com/facebook/litho/pull/963

Differential Revision: D50966209

fbshipit-source-id: d9b40b1a3895d756403d8f45314cdd9fa97702f6