Add layout params for `LazyCollection::children`
Summary:
Since we have a bunch of layout params for `LazyCollection`, whereas we're only able to set them for each single child instead of the whole collection, this diff is to add this ability for children call.

Another approach is overloading tail lambda to return a component with layout params, which is still an experimental feature([OverloadResolutionByLambdaReturnType](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-overload-resolution-by-lambda-return-type/)) and might not be a better way for now, so I decided to go with the current solution.

The usage is like:

```
// List
LazyList {
    children(friends, id = { it.id }, isSticky = { it.id == 0 }) {
      Text(it.name)
    }
  }

// Grid
LazyGrid(
        columns = 4,
    ) {
  children(
    items = (0..20),
    id = { it },
    spanSize = { if (it % 5 == 0) 4 else 1 }) {
       Text("$it")
  }
}
```

Reviewed By: astreet

Differential Revision: D51071591

fbshipit-source-id: 0bc3a07ba968a6d8d1c44831c8f83fba8eacd53e