Revert D51349204: Multisect successfully blamed "D51349204: [Litho][BE][2/x]Simplify `collectResults`" for otest failure
Summary:
This diff is reverting D51349204
D51349204: [Litho][BE][2/x]Simplify `collectResults` by pentiumao has been identified to be causing the following test failure:

Tests affected:
- [fbsource//fbandroid/instrumentation_tests/com/facebook/messaging/business/customerfeedback/view:view - BinaryYesNo_noIndicator_lightScheme (com.facebook.messaging.business.customerfeedback.view.CustomerFeedbackRatingContainerUITest)](https://www.internalfb.com/intern/test/281475078396874/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3562574
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: pentiumao

Differential Revision: D51588915

fbshipit-source-id: 563252ab964189de85b1581ae9362ca4020b5b98