Add support for keyboard focusability of links
Summary:
Lack of support for keyboard focus of links was a large gap in Litho, which was blocking users of physical keyboards from using our apps. Support for this was added by Google to the ExploreByTouchHelper class (which Litho's AccessibilityDelegate depends on) a few years back, but Litho was never updated to support this new functionality.

This adds two new annotations, `OnPerformActionForVirtualView` and `OnVirtualViewKeyboardFocusChanged`, which are required for clicking (or performing other accessibilityActions) on links, as well as updating the display of them visually when they are focused and unfocused.

Usage of these annotations have been added to the core TextSpec class, but can also be used by any higher level Litho component.

Reviewed By: adityasharat

Differential Revision: D50910356

fbshipit-source-id: f0c048313e41a9910f6ad15ab4fd8cf74db5374e