Create RecyclerBinderConfig
Summary:
Creates a `RecyclerBinderConfig` which is planned to be the single config used by both `RecyclerBinderConfiguration` (later to be renamed to `RecyclerCollectionConfiguration`) and `RecyclerBinder.Builder`.

In this diff I migrate only the `lithoViewFactory` and the `isCircular`.

On this stack I will be migrating more and more of the common configurations. At the end I will do the proper renamings for the `RecyclerBinderConfiguration`.

Reviewed By: pentiumao

Differential Revision: D51751899

fbshipit-source-id: 3e3c0377a8e8ac7606c3e6793b13bf47db0342d4