Remove layoutDiffing configuration
Summary: We are removing all code around the configuration of `layoutDiffing`. Our safe default is to enable it, and we performed QEs to force enable it on FB4A and M4A which shown neutral results (and corresponding rollouts). Therefore, we can remove all code related to it and codeship the "enable" path.

Reviewed By: zielinskimz

Differential Revision: D52017627

fbshipit-source-id: 76ccb7e1dbfc8f8bab94dccffb3930cda829a25c