Avoid javax.net.ssl.SSLHandshakeException: sun.security.validator.ValidatorException
Summary:
Recently we have been facing an issue while attempting to run the `android-setup.sh`, where it says it can't find `platforms;android-33`.

After adding `--verbose` flag, the logs indicate that there is an exception listing the packages which is ` javax.net.ssl.SSLHandshakeException: sun.security.validator.ValidatorException` .

We have been asking for support for this situation, but in parallel found  that there are other instances across Meta that faced this issue and that used the `--no_https` flag to avoid this exception (https://fburl.com/code/12z25q5j).

We are applying the same solution until we can find a more reliable option.

Reviewed By: adityasharat

Differential Revision: D52448824

fbshipit-source-id: 69154aa66d3f8d914e7c5f0f6d2b129c116b5f1b