Revert D51893192: Multisect successfully blamed "D51893192: [Litho][BE]Create RenderTreeNode with padding info" for otest failure
Summary:
This diff is reverting D51893192
D51893192: [Litho][BE]Create RenderTreeNode with padding info by pentiumao has been identified to be causing the following test failure:

Tests affected:
- [fbsource//fbandroid/instrumentation_tests/com/facebook/photos/albumcreator/ui:ui_asan - testRootComponentWithNoExtraFeatures (com.facebook.photos.albumcreator.ui.AlbumCreatorComponentsScreenshotTest)](https://www.internalfb.com/intern/test/281475078346685/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3861720
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: pentiumao

Differential Revision: D52481883

fbshipit-source-id: a8c954a197b3c52015f759a18909cacb7ca73b99