Retrieve reconciliation from ComponentsConfig
Summary: This diff is part of a series of changes aimed at improving the Litho library's configuration organization. Specifically, it removes the `isReconciliationEnabled` parameter from various classes and methods, and instead retrieves it from the `ComponentsConfiguration` object. This change is intended to simplify the code and improve consistency across the library.

Reviewed By: adityasharat, colriot

Differential Revision: D52266531

fbshipit-source-id: 960fae553416977686c5e944f05b50417ae53909