Fix showing keyboard when focus requested on appear
Summary:
This diff fixed an issue where the keyboard isn't shown when the focus was requested on TextInput component in onVisible callback.

There was some existing code that is supposed to fix race conditions in Input Method Manager but it doesn't work for some reason. I've tried to find when it stopped working, went back to Jun 2023 and the issue was already there. The IMM fix has been added in D22921269 and I've tried to checkout this commit but wasn't able to build such an old version because of some weird buck errors.

According to: https://developer.android.com/develop/ui/views/touch-and-input/keyboard-input/visibility#ShowReliably in order for keyboard to show up the window has to be focused, so in this diff I've added logic that checks if window is focused and if it's not then we're adding `OnWindowFocusChangeListener` and asking the keyboard to be shown when the window gets focus.

Reviewed By: pengj

Differential Revision: D52244988

fbshipit-source-id: 28af5981262602f4be693b8f979dcb7508d10603