Ensure we only create debug layout props once
Summary: While implementing live editting I had an issue where if you made changes  to both padding start and end you would lose the padding start changes, I noticed that this object is recreated each time and seemed to be the culprit. It seems like an oversight and should be reused. However looking for a sense check

Reviewed By: adityasharat

Differential Revision: D52515076

fbshipit-source-id: 805d5a4673c4185601f12c08dac34e14b375d1d7