Make RecyclerBinder ComponentsConfiguration non null.
Summary:
As part of the process of guaranteeing the correct propagation of configs across Litho, we want to guarantee that there are no null configuration objects anywhere in the process, as it can be a "breaking" point.

In this diff we make changes to guarantee that the `RecyclerBinder` either uses the passed `configuration` in the `RecyclerBinderConfig` or resorts to the one of the owning `Context`. This behavior was already guaranteed in practical terms due to logic in the nested `ComponentTree` creation, but this change enforces correctness and transparency.

Reviewed By: pentiumao

Differential Revision: D52599260

fbshipit-source-id: 24ac6c06339b80a4f15822caf5180854061f0e1a