Convert ContentAllocator to kotlin
Summary: This diff is converting the `ContentAllocator` class from Java to Kotlin. The changes include updating the import statements and converting the class to Kotlin syntax. The changes also include updating the `MountBehavior` class to use the new Kotlin syntax for accessing properties and methods. Finally, the `MountSpecLithoRenderUnit` class is updated to use the new Kotlin syntax for defining properties.

Reviewed By: astreet

Differential Revision: D52623591

fbshipit-source-id: 58a0d57bc51819dd85d42fa4669644d51b82c16a