Ignore broken tests (#972)
Summary:
Pull Request resolved: https://github.com/facebook/litho/pull/972

This test was one of the ones we needed to fix on Gradle to unblock the OSS release. However, after some inspection (https://www.internalfb.com/intern/test/844425049824253?ref_report_id=0 , https://www.internalfb.com/intern/test/281475075699748?ref_report_id=0), I have noticed this has been broken for a loooooooong time, and that for now we are better in unblocking ourselves and then revist this.

Reviewed By: astreet

Differential Revision: D52628227

fbshipit-source-id: b6c96fbc17f5bd4f697ff7250b9586b9f2d064f2