Move preallocation configs into `ComponentsConfig`
Summary:
This is another diff on the initiative to unify our configurations. In this diff I focus on moving the preallocation related configurations into the `ComponentsConfiguration` object.

I have renamed the APIs also a bit better, according to feedback on the Preallocation API RFC (even though they are still not in the final agreed shape - but I think that has to be done in another diff).

Reviewed By: pengj

Differential Revision: D52542146

fbshipit-source-id: 043feae9230404d842a69c775e04e122a2969059