Make host recycling configuration part of `ComponentsConfiguration`
Summary:
The goal of this diff is to move the configuration that enables host recycling from being static, to being a value inside the `ComponentsConfiguration`.

By doing this, we can actually make it configurable at the `ComponentTree` level and test it in individual surfaces (since the global launch is proving so challenging).

Reviewed By: adityasharat

Differential Revision: D52784185

fbshipit-source-id: fbf3560813199525f6792e11105da66ce81ae98b