Make HostMountContentPool preallocation behavior consistent with default item pool
Summary:
While trying to enable preallocation and host recycling on Stories I have noticed that there was an inconsistency on behavior for the `maybePreallocateContent` in the `HostComponent`.

In the default pool, we simply attempt to preallocate directly in the internal pool implementation (https://fburl.com/code/8grng0ju). However, in the host component pool we check if it could preallocate (which is not needed), because the `LayoutState` already does that pre-allocation check (https://fburl.com/code/wxwv90gi).

In this diff I make then the `HostMountContentPool` to directly call `maybePreallocate` to the internal pool.

Reviewed By: zielinskimz

Differential Revision: D52958862

fbshipit-source-id: 82e3869628bab8fb017a11b63ae4171f6034a3e0