Move KotlinSpecHelper to Model Directory
Summary:
## This Diff

Moves the `KotlinSpecHelper` to the `model` repository, next to the existing `SpecModelUtils`.

## More Context

In an effort to improve some of the build errors from Kotlin specs, I want to reuse some of the methods in the `KotlinSpecHelper`.  However, this would cause a circular dependency issue given the current paths.

Since the `SpecModelUtils` already live in `model` (and I likely want to add some of the new invocation of these functions into there), this moves the `KotlinSpecHelper` class to be in the same module.

Reviewed By: colriot

Differential Revision: D53111055

fbshipit-source-id: d813bd0422a8daa879ead9bd2427e30d78db586a