Improve Error Message for Covariance in @State and @CachedValue
Summary:
## This Diff

Add better error messaging when the issue is specifically with covariant types for `State` and `CachedValue`.

## More Context

The current error message lacks the needed information to easily resolve this issue when a Kotlin spec is written.  This is because it’s generic about types needing to match, and by all accounts, the issue is *specifically* that the types *do* look like they match, but KAPT sometimes includes covariances and other times removes it.

In short, we end up with:
```
// Source code
StateValue<List<T>>
State thing: List<T>
```

But it ends up with types:
```
List<? extends T>
List<T>
```

within Java — and those do not directly match when compared.

The basic idea here is — we can actually detect this relatively easily. and so we should start providing the information/steps for the user.

Reviewed By: colriot

Differential Revision: D53113166

fbshipit-source-id: 48cd68f78ff3592b0c507069d9e1911a4a3878a8