Simplify Some Test Setup in StateValidation
Summary:
While here, I noticed that we needed to always set all of these to empty due to how we’re mocking.  Instead of needing to do this in the majority of the tests, let’s just do that in setup (done in prior diff when adding my test cases), and then people can override it as needed.

Splitting the cleanup (deletion) out into its own diff.

Reviewed By: colriot

Differential Revision: D53113285

fbshipit-source-id: 49e0fb8c579b4f40734ab37e7dc808c52934fee4