Copies cached values from state handler
Summary: Copies cached values from state handler. The cached values were not being copied to the new state handler. This isn't a problem for the ComponentTree because component always wrote to and read cached values from the committed StateHandler. The nested tree needs to operate on the StateHandler that was create during the render phase so the cached values need to be copied.

Reviewed By: pentiumao

Differential Revision: D52943173

fbshipit-source-id: 8f3fbf2be2fd35d497af6ab4dacbfa855fcac0a8